+++ SuperMarketApp/WebApplication/Interfaces/ICalculateProductPrice.cs
@@ -1,9 +1,10 @@
 ﻿using Service.Models;
+using SuperMarketApp.Service.Models;
 
 namespace Service.Interfaces
 {
     public interface ICalculateProductPrice
     {
-        double Calculate(Product product);
+        decimal Calculate(Product product);
     }
 }
